Sweet alert when creating, deleting or editing a task

+++ src/components/display/ListDropdown.jsx
@@ -5,13 +5,14 @@ import {
 } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { 
+import {
     faPenToSquare,
     faTrashCan,
 } from '@fortawesome/free-regular-svg-icons';
-import { 
+import {
     faEllipsisVertical
 } from '@fortawesome/free-solid-svg-icons';
+import Swal from 'sweetalert2';
 import { getFirestore, deleteDoc, doc } from 'firebase/firestore';
 import appFirebase from '../../firebase';
 
@@ -19,7 +20,20 @@ const db = getFirestore(appFirebase);
 
 const ListDropdown = ({ taskId }) => {
     const deleteTask = async (id) => {
-        await deleteDoc(doc(db, "tasks", id))
+        Swal.fire({
+            icon: 'warning',
+            title: 'Do you want to delete this task?',
+            text: "You won't be able to revert this!",
+            showCancelButton: true,
+            confirmButtonText: 'Delete',
+            confirmButtonColor: '#d33',
+        }).then((result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+                deleteDoc(doc(db, "tasks", id))
+                Swal.fire('Task deleted!', '', 'success')
+            }
+        })
     }
 
     return (
@@ -30,7 +44,7 @@ const ListDropdown = ({ taskId }) => {
 
             <Dropdown.Menu>
                 <Dropdown.Item as={Link} to={`/tasks/${taskId}`}>
-                        <FontAwesomeIcon icon={faPenToSquare} /> Edit
+                    <FontAwesomeIcon icon={faPenToSquare} /> Edit
                 </Dropdown.Item>
                 <Dropdown.Item as={Button} onClick={() => deleteTask(taskId)}>
                     <FontAwesomeIcon icon={faTrashCan} /> Delete
